requirements.txt
loguru==0.7.3
paho-mqtt==1.6.1
